ADD: post controller, route

+++ src/controllers/post.controller.ts
@@ -3,6 +3,7 @@ import { asyncHandler } from "../utils/async-handler";
 import { ApiError } from "../utils/api-error";
 import { Post } from "../models/post.model";
 import { ApiResponse } from "../utils/api-response";
+import { PostStatusEnum, UserRolesEnum } from "../utils/constants";
 
 const createPost = asyncHandler(async (req: Request, res: Response) => {
   const { title, content, category } = req.body;
@@ -59,7 +60,15 @@ const getPostById = asyncHandler(async (req: Request, res: Response) => {
     .json(new ApiResponse(200, `Post fetched successfully`, post));
 });
 
-const getAllPost = asyncHandler(async (req: Request, res: Response) => {});
+const getAllPost = asyncHandler(async (req: Request, res: Response) => {
+  const post = await Post.find({ status: PostStatusEnum.APPROVE });
+
+  if (!post) {
+    throw new ApiError(404, "Approve post not available");
+  }
+
+  return res.status(200).json(new ApiResponse(200, "Fetched all post", post));
+});
 
 const deletePostById = asyncHandler(async (req: Request, res: Response) => {
   const { pid } = req.params;
@@ -72,6 +81,13 @@ const deletePostById = asyncHandler(async (req: Request, res: Response) => {
 
   if (!post) {
     throw new ApiError(404, "This blog post not found");
+  }
+
+  if (post.status !== PostStatusEnum.REJECT) {
+    throw new ApiError(
+      403,
+      "ACCESS DENIED, You can't delete due to author doesn't reject your post",
+    );
   }
 
   const delPost = await Post.findByIdAndDelete(post?._id);
@@ -99,7 +115,14 @@ const updatePostById = asyncHandler(async (req: Request, res: Response) => {
   const post = await Post.findById(pid);
 
   if (!post) {
-    throw new ApiError(404, "This blog post not found");
+    throw new ApiError(404, "post not found");
+  }
+
+  if (post.status !== PostStatusEnum.REJECT) {
+    throw new ApiError(
+      403,
+      "ACCESS DENIED, You can't delete due to author doesn't reject your post",
+    );
   }
 
   if (!title || !content || !category) {
@@ -138,4 +161,97 @@ const updatePostById = asyncHandler(async (req: Request, res: Response) => {
     .json(new ApiResponse(200, `Post delete successfully`, updatePost));
 });
 
-export { createPost, getPostById, getAllPost, deletePostById, updatePostById };
+const getAllPendingPostByAdmin = asyncHandler(
+  async (req: Request, res: Response) => {
+    if (!req.user) {
+      throw new ApiError(404, "Requested user not found");
+    }
+
+    const user = req.user;
+
+    if (user.role !== UserRolesEnum.ADMIN) {
+      throw new ApiError(403, "ERROR: Permission and Access Denied");
+    }
+
+    const post = await Post.find({ status: PostStatusEnum.PENDING });
+
+    if (!post) {
+      throw new ApiError(404, "No pending post available");
+    }
+
+    return res
+      .status(200)
+      .json(new ApiResponse(200, "Fetched all pending post", post));
+  },
+);
+
+const approvePostByAdmin = asyncHandler(async (req: Request, res: Response) => {
+  if (!req.user) {
+    throw new ApiError(404, "Requested user not found");
+  }
+
+  const user = req.user;
+  const { pid } = req.params;
+
+  if (user.role !== UserRolesEnum.ADMIN) {
+    throw new ApiError(403, "ERROR: Permission and Access Denied");
+  }
+
+  if (!pid) {
+    throw new ApiError(404, "Requested postId not found");
+  }
+
+  const post = await Post.findById(pid);
+
+  if (!post) {
+    throw new ApiError(404, "post doesn't exist");
+  }
+
+  post.status = PostStatusEnum.APPROVE;
+  post.save({ validateBeforeSave: false });
+
+  return res
+    .status(200)
+    .json(new ApiResponse(200, "Fetched all approve post", post));
+});
+
+const rejectPostByAdmin = asyncHandler(async (req: Request, res: Response) => {
+  if (!req.user) {
+    throw new ApiError(404, "Requested user not found");
+  }
+
+  const user = req.user;
+  const { pid } = req.params;
+
+  if (user.role !== UserRolesEnum.ADMIN) {
+    throw new ApiError(403, "ERROR: Permission and Access Denied");
+  }
+
+  if (!pid) {
+    throw new ApiError(404, "Requested postId not found");
+  }
+
+  const post = await Post.findById(pid);
+
+  if (!post) {
+    throw new ApiError(404, "post doesn't exist");
+  }
+
+  post.status = PostStatusEnum.REJECT;
+  post.save({ validateBeforeSave: false });
+
+  return res
+    .status(200)
+    .json(new ApiResponse(200, "Fetched all reject post", post));
+});
+
+export {
+  createPost,
+  getPostById,
+  getAllPost,
+  deletePostById,
+  updatePostById,
+  approvePostByAdmin,
+  getAllPendingPostByAdmin,
+  rejectPostByAdmin,
+};
